multi module에서 단일 모듈로 변경
