+++ README.md
@@ -1,8 +1,2 @@
 # GPTBoard
 
-## 멀티모듈 구조
-
-- `app`: 애플리케이션 실행을 위한 엔트리포인트를 담당하는 모듈
-- `web`: 웹 애플리케이션을 위한 모듈
-- `domain`: 비즈니스 로직을 담당하는 도메인 모듈
-- `common`: 공통으로 사용되는 코드, 라이브러리 등을 포함하는 모듈
