--- app/Dockerfile
@@ -1,5 +1,0 @@
-FROM openjdk:11
-ARG JAR_FILE=build/libs/app.jar
-COPY ${JAR_FILE} ./app.jar
-ENV TZ=Asia/Seoul
-ENTRYPOINT ["java", "-jar", "./app.jar"]
