--- app/build.gradle
@@ -1,40 +1,0 @@
-plugins {
-    id 'java'
-    id 'org.springframework.boot' version '2.7.9'
-    id 'io.spring.dependency-management' version '1.0.15.RELEASE'
-}
-
-group = 'dev.be'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '11'
-
-bootJar {
-    archiveFileName = 'app.jar'
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-
-    // Lombok
-    compileOnly 'org.projectlombok:lombok'
-    annotationProcessor 'org.projectlombok:lombok'
-
-    // mysql
-    runtimeOnly 'com.mysql:mysql-connector-j'
-
-    // multi module
-    implementation project(':common')
-    implementation project(':domain')
-    implementation project(':web')
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
-
-tasks.register("prepareKotlinBuildScriptModel") {}
