--- app/src/main/resources/application-local.yml
@@ -1,16 +1,0 @@
-spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    username: root
-    url: jdbc:mysql://localhost:3306/gpt-board?characterEncoding=UTF-8&serverTimeZone=Asia/Seoul
-    password: gpt-board-1234
-
-  jpa:
-    hibernate:
-      ddl-auto: create
-    show-sql: true
-    properties:
-      hibernate:
-        format_sql: true
-        default_batch_fetch_size: 100
-    open-in-view: false
