--- common/docker/docker-compose-local.yml
@@ -1,17 +1,0 @@
-version: "3"
-
-services:
-  database:
-    image: mysql:8.0
-    container_name: gpt-board
-    ports:
-      - "3306:3306"
-    environment:
-      - MYSQL_DATABASE=gpt-board
-      - MYSQL_ROOT_PASSWORD=gpt-board-1234
-      - TZ=Asia/Seoul
-    command:
-      - --character-set-server=utf8mb4
-      - --collation-server=utf8mb4_unicode_ci
-    platform: linux/amd64
-
