--- common/src/main/java/dev/be/common/config/JpaConfig.java
@@ -1,9 +1,0 @@
-package dev.be.common.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-}
