+++ settings.gradle
@@ -1,6 +1,1 @@
 rootProject.name = 'GPTBoard'
-
-include 'app'
-include 'common'
-include 'domain'
-include 'web'
