+++ src/main/java/dev/be/gptboard/GptBoardApplication.java
@@ -1,0 +1,12 @@
+package dev.be.gptboard;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GptBoardApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(GptBoardApplication.class, args);
+    }
+}
