+++ src/main/resources/data.sql
@@ -1,0 +1,9 @@
+insert into member (id, nickname, password, email, created_at, created_by, modified_at, modified_by) values
+                                                                                                        (1, 'hankwanjin', '1234', 'hankwanjin@email.com', '2023-03-11 11:11:11', 'hankwanjin', '2023-03-12 11:11:11', 'hankwanjin');
+insert into article (id, member_id, title, content, created_at, created_by, modified_at, modified_by) values
+                                                                                                        (1, 1, 'title-test-1', 'content-test-1', '2023-03-11 11:11:11', 'hankwanjin', '2023-03-12 11:11:11', 'hankwanjin'),
+                                                                                                        (2, 1, 'title-test-2', 'content-test-2', '2023-03-11 11:11:12', 'hankwanjin', '2023-03-12 11:11:12', 'hankwanjin'),
+                                                                                                        (3, 1, 'title-test-3', 'content-test-3', '2023-03-11 11:11:13', 'hankwanjin', '2023-03-12 11:11:13', 'hankwanjin');
+insert into article_comment (id, article_id, member_id, parent_comment_id, content, created_at, created_by, modified_at, modified_by) values
+                                                                                                                                          (1, 1, 1, null, 'test-content-1', '2023-03-11 11:11:11', 'hankwanjin', '2023-03-12 11:11:11', 'hankwanjin'),
+                                                                                                                                          (2, 1, 1,    1, 'test-content-2', '2023-03-11 11:11:11', 'hankwanjin', '2023-03-12 11:11:11', 'hankwanjin');
