+++ src/test/java/dev/be/gptboard/GptBoardApplicationTests.java
@@ -1,0 +1,11 @@
+package dev.be.gptboard;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Disabled;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@Disabled
+@SpringBootTest
+class GptBoardApplicationTests {
+}
