--- web/build.gradle
@@ -1,34 +1,0 @@
-plugins {
-    id 'java'
-    id 'org.springframework.boot' version '2.7.9'
-    id 'io.spring.dependency-management' version '1.0.15.RELEASE'
-    id 'java-library'
-}
-
-group = 'dev.be'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '11'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-
-    // Lombok
-    compileOnly 'org.projectlombok:lombok'
-    annotationProcessor 'org.projectlombok:lombok'
-
-    implementation project(':common')
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
-
-tasks.bootJar { enabled = false }
-tasks.jar { enabled = true }
-
-tasks.register("prepareKotlinBuildScriptModel") {}
